tests/wpt/tests/css/css-images/animation/image-no-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<link rel=help href="https://drafts.csswg.org/css-transitions-2/#transition-property-property">
<link rel=help href="https://drafts.csswg.org/css-images/#the-image-rendering">
<link rel=help href="https://drafts.csswg.org/css-images/#the-image-orientation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
test_no_interpolation({
  property: 'image-orientation',
  from: 'initial',
  to: 'none'
});

test_no_interpolation({
  property: 'image-rendering',
  from: 'initial',
  to: 'pixelated'
});
</script>
